シーンマネージャを実装したときに当たり判定の不具合が発生したので修正

+++ DirectXGame/Scene/GamePlayScene.h
@@ -149,7 +149,7 @@ private://メンバ変数
 	//ボスの数
 	int bossNum_;
 	//墜落演出の墜落量
-	float gameOverNum_;
+	int gameOverNum_;
 
 	//フラグ
 	//ボス戦に入ったか
